+++ devtools/devtools.js
@@ -1,0 +1,5 @@
+chrome.devtools.panels.create(
+  "MinTool",
+  "",
+  "services/cors-bypass/panel.html",
+);